requirements.txt
Django>=3.2
psycopg2-binary>=2.9.1
minio>=7.0.0
djangorestframework>=3.12.0
django-filter>=21.1
drf-yasg>=1.20.0
redis>=4.3.4
django-redis>=5.2.0